Tweaked code for finding image ID, and added version.
A version number was added for easy identification of the script.
The regex code for finding the IID was improved so that it would correctly capture IID's containing special characters such as ^.
The log text for when an IID was not found was improved. Additionally, a TODO note was left indicating possible future improvements.